Replace harbor symlink if it already exists
There was an issue where if you ran nix develop, then logged out,
then ran nix develop again...you'd get an error cos it was trying to
symlink a file that was already existing as a symlink.  The -n flag
combined with the -f flag will fix this, doing an overwrite of the whole
directory as needed.